refactor: common파일을 제거하여 폴더구조 수정
