+++ apps/client/src/pages/search/components/search-result/notice-section.css.ts
