+++ apps/client/src/pages/search/components/search-result/notice-section.tsx
