+++ apps/client/src/pages/search/components/search-result/title.css.ts
